bot/requirements.txt
aiogram==3.18.0
APScheduler==3.10.4
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0